requirements.txt
flask
psycopg2
plotly.graph_objs as go
plotly.utils
json
numpy
decimal
csv
os
pandas
config
regex
